web/requirements.txt
# BASE
Django==3.0.5
djangorestframework==3.11.0
djangorestframework-gis==0.15
djangorestframework-simplejwt==4.4.0
django-redis==4.3.0
django_excel==0.0.10
django_requestlogs==0.2.1
PyJWT==1.6.4
redis==3.2.0
celery==4.4.0
psycopg2-binary==2.8.4
geocoder==1.38.1
django-ipware==2.1.0
django-geojson==3.0.0
trimesh=3.6.19
# PROD
gunicorn==19.5.0

# DEV
django-debug-toolbar==2.1
django-extensions==2.2.5
coverage==5.0.3
ipdb==0.12.3
pylint==2.4.4
pylint-django==2.0.13